Components and Views Refactored
