+++ sources/packages/web/src/components/partial-view/student/NavBar.vue
@@ -2,7 +2,7 @@
   <Menubar :model="menuItems" class="gov-header">
     <template #start>
       <div class="p-menubar-start">
-        <img alt="logo" src="../assets/images/bc_logo.svg" height="40" />
+        <img alt="logo" src="../../../assets/images/bc_logo.svg" height="40" />
         <span class="title">Student Aid</span>
       </div>
     </template>
@@ -29,7 +29,7 @@
 
 <script lang="ts">
 import { computed } from "vue";
-import { AppConfigService } from "../../services/AppConfigService";
+import { AppConfigService } from "../../../services/AppConfigService";
 export default {
   setup() {
     const menuItems: any = [];
