+++ sources/packages/web/tests/unit/components/fa-application/Section.spec.ts
@@ -1,5 +1,5 @@
 import { mount } from "@vue/test-utils";
-import Section from "../../../../src/components/shared/Section.vue";
+import Section from "../../../../src/components/generic/Section.vue";
 
 const sectionTitleSelector = "#sectionTitle";
